+++ src/main/java/baseball/app/game/Player.java
@@ -2,7 +2,7 @@ package baseball.app.game;
 
 import baseball.app.ball.Balls;
 import baseball.app.ball.generator.BallsGenerator;
-import baseball.app.printer.GameView;
+import baseball.app.view.GameView;
 import camp.nextstep.edu.missionutils.Console;
 
 public class Player {
